pyproject.toml
[project]
name = "csv-to-yaml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pyyaml>=6.0.3",
    "jsonschema>=4.22.0",
    "pytest>=8.3.0",
    "ruamel.yaml>=0.18.0",
]
